fix(ui): unify queued chat ownership across reconnects
