+++ ui/src/lib/chat/outbox-store.test.ts
@@ -2,6 +2,7 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import { createStorageMock } from "../../test-helpers/storage.ts";
 import {
+  listStoredChatOutboxes,
   resolveStoredChatOutboxScope,
   storedChatOutboxScopeKey,
   subscribeStoredChatOutboxChanges,
@@ -143,6 +144,21 @@ describe("stored outbox summaries", () => {
       resolveStoredChatOutboxScope(offlineState, "main"),
     );
     expect(summary.countsByScope.get(sidebarScopeKey)).toBe(1);
+    expect(listStoredChatOutboxes(offlineState)).toEqual([
+      {
+        sessionKey: "global",
+        agentId: "work",
+        queue: [
+          {
+            id: "queued",
+            text: "queued",
+            createdAt: 1,
+            sessionKey: "global",
+            agentId: "work",
+          },
+        ],
+      },
+    ]);
   });
 
   it("rejects a v2 store owned by another gateway", () => {
@@ -228,5 +244,51 @@ describe("stored outbox summaries", () => {
     expect(summary.total).toBe(2);
     expect(summary.countsByScope.get(storedChatOutboxScopeKey({ sessionKey: "thread-a" }))).toBe(1);
     expect(summary.countsByScope.get(storedChatOutboxScopeKey({ sessionKey: "thread-b" }))).toBe(1);
+  });
+
+  it("derives badges and replay from the same migrated durable queue", () => {
+    const gatewayUrl = "ws://gateway.test/control";
+    const legacyKey = `openclaw.control.chatComposer.v1:${encodeURIComponent(gatewayUrl)}`;
+    sessionStorage.setItem(
+      legacyKey,
+      JSON.stringify({
+        version: 1,
+        sessions: {
+          "main\u0000agent:previous": {
+            queue: [
+              { id: "removed", text: "removed", createdAt: 1 },
+              { id: "shared", text: "older", createdAt: 2 },
+            ],
+            removedQueueItemIds: ["removed"],
+            updatedAt: 2,
+          },
+          "global\u0000agent:work": {
+            queue: [{ id: "shared", text: "newer", createdAt: 3 }],
+            updatedAt: 3,
+          },
+        },
+      }),
+    );
+    const state = {
+      settings: { gatewayUrl },
+      assistantAgentId: "work",
+      agentsList: { defaultId: "work", mainKey: "main" },
+    };
+
+    const summary = summarizeStoredChatOutboxes(state);
+    const outboxes = listStoredChatOutboxes(state);
+
+    expect(summary.total).toBe(1);
+    expect(summary.countsByScope.get(storedChatOutboxScopeKey(outboxes[0]!))).toBe(1);
+    expect(outboxes[0]?.queue).toEqual([
+      {
+        id: "shared",
+        text: "newer",
+        createdAt: 3,
+        sessionKey: "global",
+        agentId: "work",
+      },
+    ]);
+    expect(sessionStorage.getItem(legacyKey)).toBeNull();
   });
 });
